Revert "Make API controllers log to BigQuery"
