+++ app/controllers/candidate_api/candidates_controller.rb
@@ -1,7 +1,11 @@
 module CandidateAPI
-  class CandidatesController < ApplicationAPIController
+  class CandidatesController < ActionController::API
     include ServiceAPIUserAuthentication
+    include RemoveBrowserOnlyHeaders
     include Pagy::Backend
+
+    rescue_from ActionController::ParameterMissing, with: :parameter_missing
+    rescue_from ParameterInvalid, with: :parameter_invalid
 
     # Makes PG::QueryCanceled statement timeout errors appear in Skylight
     # against the controller action that triggered them
@@ -16,6 +20,15 @@ module CandidateAPI
 
     def index
       render json: { data: serialized_candidates }
+    end
+
+    def parameter_missing(e)
+      error_message = e.message.split("\n").first
+      render json: { errors: [{ error: 'ParameterMissing', message: error_message }] }, status: :unprocessable_entity
+    end
+
+    def parameter_invalid(e)
+      render json: { errors: [{ error: 'ParameterInvalid', message: e }] }, status: :unprocessable_entity
     end
 
     def statement_timeout
