+++ app/controllers/register_api/applications_controller.rb
@@ -1,7 +1,9 @@
 module RegisterAPI
-  class ApplicationsController < ApplicationAPIController
+  class ApplicationsController < ActionController::API
     include ServiceAPIUserAuthentication
+    include RemoveBrowserOnlyHeaders
 
+    rescue_from ActionController::ParameterMissing, with: :parameter_missing
     rescue_from ParameterInvalid, with: :parameter_invalid
 
     # Makes PG::QueryCanceled statement timeout errors appear in Skylight
@@ -11,6 +13,15 @@ module RegisterAPI
 
     def index
       render json: { data: serialized_application_choices }
+    end
+
+    def parameter_missing(e)
+      error_message = e.message.split("\n").first
+      render json: { errors: [{ error: 'ParameterMissing', message: error_message }] }, status: :unprocessable_entity
+    end
+
+    def parameter_invalid(e)
+      render json: { errors: [{ error: 'ParameterInvalid', message: e }] }, status: :unprocessable_entity
     end
 
     def statement_timeout
