--- lib/rubocop/cop/application_api_controller.rb
@@ -1,13 +1,0 @@
-module RuboCop
-  module Cop
-    class ApplicationAPIController < Base
-      extend AutoCorrector
-
-      MSG = 'API controllers should subclass `ApplicationAPIController`.'.freeze
-      SUPERCLASS = 'ApplicationAPIController'.freeze
-      BASE_PATTERN = '(const (const nil? :ActionController) :API)'.freeze
-
-      include RuboCop::Cop::EnforceSuperclass
-    end
-  end
-end
